+++ src/Page/MangeOrder/styles.ts
@@ -1,0 +1,5 @@
+import styled from "styled-components";
+
+export const Container = styled.div``;
+export const OrderStateHeader = styled.div``;
+export const OrderStateContainer = styled.div``;
